Refactoring of CSS/JS loading and finish layout of membership-overview
CSS/JS loading: All register_style/register_script calls are collected
in class-ms-controller.php to give us an overview of what files are used
by the plugin. In the controllers we only use the enqueue function to
load the files.
Also most layout-issues in other pages that were caused by recent changes
are fixed with this update.