Implement the finished tag-select component for Categories and CPT
A new component-type "tag_select" is available in the HTML helper class.
This component will create several input elements:

A select list to choose an item, a button to add the selected item to
the tag list and the tag list itself. This component also needs some
javascript and CSS to work (provided in ms-admin.js and ms-admin.css)