Resolve collission of comment-permission for multiple active memberships
When a user is member of more than one membership then these memberships
can define different comment conditions. When this was the case the
plugin now uses the MOST GENEROUS permission.

E.g. If all is set to "no comments", but one membership uses
"read and write" then the user has full access to comments.