Allow shortcodes to be replaced in multiple calls to the_content
Until now we had a shortcode-memory that kept track if we already
inserted a certain shortcode to the page. This makes sense when we don't
want to insert two login-forms on the same page. However, Jetpack uses
the_content already in the page-header which in turn set all shortcodes
on the page to "already inserted". This caused problems esp. during the
registration process