Solve compatibility issue with PopUp, Custom Sidebars and Google Maps
The shared component "WDev()" has changed a lot and is not compatible
with the other plugins. To solve the incompatibility we created a new
version of WDev() - now called lib2() - that can be used side by side
with the older version.

Other plugins use WDev() and do not know anything about the new library.
Protected Content uses lib2() and ignores the old WDev() component.