Critical bug fix caused by timezone offset
Some users were complaining that all Subscriptions were deactivated over night.
We could not reproduce, because it turns out that the issue was caused by a
function using the local timezone (while current_date uses the GMT-zone)

To solve this we force empty end-dates to be interpreted as 2999-12-31 in the
subtract_date function.