Fix the pre_get_posts handler to prevent incorrect redirects
Because of new output buffering in media-protection sometimes the page was
redirected to front page without a reason. Turned out that the payment gateway
handler was always executed when a WP_Query() was created. The gateway handler
tried to redirect to front page but usually failed because headers were already
sent.
Fixed it to only process the main WP Query