Simplify trial invoicing: Create only one invoice instead of two invoices
Old logic created 2 (almost identical) invoices when user started a trial
membership. The first invoice only specified the "is-trial" flag, while the
second invoice was the actual invoice that was displayed to the user.

New logic is to have only one invoice that also has a "includes-trial" flag
which specifies that this is the first invoice and is also consuming the trial
ability for this membership.