Improve some singleton-caching logic to avoid problems found while testing
Problem was:
When a new invoice or subscription is created it was NOT a singleton by an
uncached object. Once this new invoice was saved it will get an ID (e.g. "123")
and another function tried to access invoice with ID "123" a new singleton for
invoice 123 was created instead of re-using the existing object.

As a result we had 2 invoices (one singleton and one non-singleton) that were
not in sync and therefore had different property values.