Move check for MS_PAYMENT_DELAY into get_remaining_days() calculation
This way the flag is included in all M2 calculations. In this case we also use
the remaining_days() value for the status-change logic to decide if a
subscription should be set to expired or not.

Note: In line 2755 we have check for "<= 0" even though the user still has full
access on day 0; but this condition only tries to process next payment, it does
not expire the subscription.